# Measure peak RAM in benchmarks